+++ frontend/src/component/insights/InsightsCharts.tsx
@@ -1,25 +1,8 @@
-import type { FC, PropsWithChildren, ReactNode } from 'react';
+import type { FC } from 'react';
 import { Box, Paper, styled } from '@mui/material';
-import { UserStats } from './componentsStat/UserStats/UserStats.tsx';
-import { UsersChart } from './componentsChart/UsersChart/UsersChart.tsx';
-import { UsersPerProjectChart } from './componentsChart/UsersPerProjectChart/UsersPerProjectChart.tsx';
-import { FlagStats } from './componentsStat/FlagStats/FlagStats.tsx';
-import { FlagsChart } from './componentsChart/FlagsChart/FlagsChart.tsx';
-import { FlagsProjectChart } from './componentsChart/FlagsProjectChart/FlagsProjectChart.tsx';
-import { HealthStats } from './componentsStat/HealthStats/HealthStats.tsx';
-import { ProjectHealthChart } from './componentsChart/ProjectHealthChart/ProjectHealthChart.tsx';
-import { MetricsSummaryChart } from './componentsChart/MetricsSummaryChart/MetricsSummaryChart.tsx';
-import { UpdatesPerEnvironmentTypeChart } from './componentsChart/UpdatesPerEnvironmentTypeChart/UpdatesPerEnvironmentTypeChart.tsx';
-import { allOption } from 'component/common/ProjectSelect/ProjectSelect';
-import useUiConfig from 'hooks/api/getters/useUiConfig/useUiConfig';
-import { WidgetTitle } from './components/WidgetTitle/WidgetTitle.tsx';
-import { format, subMonths } from 'date-fns';
-import { useInsights } from 'hooks/api/getters/useInsights/useInsights.ts';
-import { usePersistentTableState } from 'hooks/usePersistentTableState.ts';
-import { withDefault } from 'use-query-params';
-import { FilterItemParam } from 'utils/serializeQueryParams.ts';
-import { useInsightsData } from './hooks/useInsightsData.ts';
-import { InsightsFilters } from './InsightsFilters.tsx';
+import { LifecycleInsights } from './sections/LifecycleInsights.tsx';
+import { PerformanceInsights } from './sections/PerformanceInsights.tsx';
+import { UserInsights } from './sections/UserInsights.tsx';
 
 const StyledContainer = styled(Box)(({ theme }) => ({
     display: 'flex',
@@ -65,38 +48,6 @@ const StyledChartContainer = styled(Box)(({ theme }) => ({
     padding: theme.spacing(3),
 }));
 
-const StyledSection = styled('section')(({ theme }) => ({
-    display: 'flex',
-    flexDirection: 'column',
-    gap: theme.spacing(2),
-}));
-
-const SectionTitleRow = styled('div')(({ theme }) => ({
-    position: 'sticky',
-    top: 0,
-    zIndex: theme.zIndex.sticky,
-    paddingBlock: theme.spacing(2),
-    background: theme.palette.background.application,
-    transition: 'padding 0.3s ease',
-    display: 'flex',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    flexFlow: 'row wrap',
-    rowGap: theme.spacing(2),
-}));
-
-const Section: FC<
-    PropsWithChildren<{ title: string; filters?: ReactNode }>
-> = ({ title, children, filters: HeaderActions }) => (
-    <StyledSection>
-        <SectionTitleRow>
-            <h2>{title}</h2>
-            {HeaderActions}
-        </SectionTitleRow>
-        {children}
-    </StyledSection>
-);
-
 export const InsightsCharts: FC<{}> = () => {
     return (
         <StyledContainer>
@@ -104,291 +55,5 @@ export const InsightsCharts: FC<{}> = () => {
             <PerformanceInsights />
             <UserInsights />
         </StyledContainer>
-    );
-};
-
-const StyledInsightsFilters = styled(InsightsFilters)({
-    padding: 0,
-});
-
-const LifecycleInsights: FC<{}> = () => {
-    const statePrefix = 'lifecycle-';
-    const stateConfig = {
-        [`${statePrefix}project`]: FilterItemParam,
-    };
-    const [state, setState] = usePersistentTableState('insights', stateConfig);
-
-    return (
-        <Section
-            title='Flags lifecycle currently'
-            filters={
-                <StyledInsightsFilters
-                    state={state}
-                    onChange={setState}
-                    filterNamePrefix={statePrefix}
-                />
-            }
-        />
-    );
-};
-
-const PerformanceInsights: FC<{}> = () => {
-    const statePrefix = 'performance-';
-    const stateConfig = {
-        [`${statePrefix}project`]: FilterItemParam,
-        [`${statePrefix}from`]: withDefault(FilterItemParam, {
-            values: [format(subMonths(new Date(), 1), 'yyyy-MM-dd')],
-            operator: 'IS',
-        }),
-        [`${statePrefix}to`]: withDefault(FilterItemParam, {
-            values: [format(new Date(), 'yyyy-MM-dd')],
-            operator: 'IS',
-        }),
-    };
-    const [state, setState] = usePersistentTableState('insights', stateConfig, [
-        'performance-from',
-        'performance-to',
-    ]);
-
-    const { insights, loading } = useInsights(
-        state[`${statePrefix}from`]?.values[0],
-        state[`${statePrefix}to`]?.values[0],
-    );
-
-    const projects = state[`${statePrefix}project`]?.values ?? [allOption.id];
-
-    const showAllProjects = projects[0] === allOption.id;
-    const {
-        flagTrends,
-        summary,
-        groupedProjectsData,
-        userTrends,
-        groupedMetricsData,
-        allMetricsDatapoints,
-        environmentTypeTrends,
-    } = useInsightsData(insights, projects);
-
-    const { isEnterprise } = useUiConfig();
-    const lastUserTrend = userTrends[userTrends.length - 1];
-    const usersTotal = lastUserTrend?.total ?? 0;
-    const lastFlagTrend = flagTrends[flagTrends.length - 1];
-    const flagsTotal = lastFlagTrend?.total ?? 0;
-
-    function getFlagsPerUser(flagsTotal: number, usersTotal: number) {
-        const flagsPerUserCalculation = flagsTotal / usersTotal;
-        return Number.isNaN(flagsPerUserCalculation)
-            ? 'N/A'
-            : flagsPerUserCalculation.toFixed(2);
-    }
-
-    return (
-        <Section
-            title='Performance insights'
-            filters={
-                <StyledInsightsFilters
-                    state={state}
-                    onChange={setState}
-                    filterNamePrefix={statePrefix}
-                />
-            }
-        >
-            {showAllProjects ? (
-                <StyledWidget>
-                    <StyledWidgetStats width={275}>
-                        <WidgetTitle title='Flags' />
-                        <FlagStats
-                            count={flagsTotal}
-                            flagsPerUser={getFlagsPerUser(
-                                flagsTotal,
-                                usersTotal,
-                            )}
-                            isLoading={loading}
-                        />
-                    </StyledWidgetStats>
-                    <StyledChartContainer>
-                        <FlagsChart
-                            flagTrends={flagTrends}
-                            isLoading={loading}
-                        />
-                    </StyledChartContainer>
-                </StyledWidget>
-            ) : (
-                <StyledWidget>
-                    <StyledWidgetStats width={275}>
-                        <WidgetTitle title='Flags' />
-                        <FlagStats
-                            count={summary.total}
-                            flagsPerUser={''}
-                            isLoading={loading}
-                        />
-                    </StyledWidgetStats>
-                    <StyledChartContainer>
-                        <FlagsProjectChart
-                            projectFlagTrends={groupedProjectsData}
-                            isLoading={loading}
-                        />
-                    </StyledChartContainer>
-                </StyledWidget>
-            )}
-            {isEnterprise() ? (
-                <StyledWidget>
-                    <StyledWidgetStats width={350} padding={0}>
-                        <HealthStats
-                            value={summary.averageHealth}
-                            healthy={summary.active}
-                            stale={summary.stale}
-                            potentiallyStale={summary.potentiallyStale}
-                            title={
-                                <WidgetTitle
-                                    title='Health'
-                                    tooltip={
-                                        'Percentage of flags that are not stale or potentially stale.'
-                                    }
-                                />
-                            }
-                        />
-                    </StyledWidgetStats>
-                    <StyledChartContainer>
-                        <ProjectHealthChart
-                            projectFlagTrends={groupedProjectsData}
-                            isAggregate={showAllProjects}
-                            isLoading={loading}
-                        />
-                    </StyledChartContainer>
-                </StyledWidget>
-            ) : null}
-            {isEnterprise() ? (
-                <>
-                    <StyledWidget>
-                        <StyledWidgetContent>
-                            <WidgetTitle
-                                title='Flag evaluation metrics'
-                                tooltip='Summary of all flag evaluations reported by SDKs.'
-                            />
-                            <StyledChartContainer>
-                                <MetricsSummaryChart
-                                    metricsSummaryTrends={groupedMetricsData}
-                                    allDatapointsSorted={allMetricsDatapoints}
-                                    isAggregate={showAllProjects}
-                                    isLoading={loading}
-                                />
-                            </StyledChartContainer>
-                        </StyledWidgetContent>
-                    </StyledWidget>
-                    <StyledWidget>
-                        <StyledWidgetContent>
-                            <WidgetTitle
-                                title='Updates per environment type'
-                                tooltip='Summary of all configuration updates per environment type.'
-                            />
-                            <UpdatesPerEnvironmentTypeChart
-                                environmentTypeTrends={environmentTypeTrends}
-                                isLoading={loading}
-                            />
-                        </StyledWidgetContent>
-                    </StyledWidget>
-                </>
-            ) : null}
-        </Section>
-    );
-};
-
-const UserInsights: FC<{}> = () => {
-    const statePrefix = 'users-';
-    const stateConfig = {
-        [`${statePrefix}project`]: FilterItemParam,
-        [`${statePrefix}from`]: withDefault(FilterItemParam, {
-            values: [format(subMonths(new Date(), 1), 'yyyy-MM-dd')],
-            operator: 'IS',
-        }),
-        [`${statePrefix}to`]: withDefault(FilterItemParam, {
-            values: [format(new Date(), 'yyyy-MM-dd')],
-            operator: 'IS',
-        }),
-    };
-    const [state, setState] = usePersistentTableState(
-        'insights-users',
-        stateConfig,
-        ['users-from', 'users-to'],
-    );
-
-    const { insights, loading } = useInsights(
-        state['users-from']?.values[0],
-        state['users-to']?.values[0],
-    );
-
-    const projects = state['users-project']?.values ?? [allOption.id];
-
-    const showAllProjects = projects[0] === allOption.id;
-    const { summary, groupedProjectsData, userTrends } = useInsightsData(
-        insights,
-        projects,
-    );
-
-    const lastUserTrend = userTrends[userTrends.length - 1];
-    const usersTotal = lastUserTrend?.total ?? 0;
-    const usersActive = lastUserTrend?.active ?? 0;
-    const usersInactive = lastUserTrend?.inactive ?? 0;
-
-    const isOneProjectSelected = projects.length === 1;
-
-    return (
-        <Section
-            title='User insights'
-            filters={
-                <StyledInsightsFilters
-                    state={state}
-                    onChange={setState}
-                    filterNamePrefix={statePrefix}
-                />
-            }
-        >
-            {showAllProjects ? (
-                <StyledWidget>
-                    <StyledWidgetStats>
-                        <WidgetTitle title='Total users' />
-                        <UserStats
-                            count={usersTotal}
-                            active={usersActive}
-                            inactive={usersInactive}
-                            isLoading={loading}
-                        />
-                    </StyledWidgetStats>
-                    <StyledChartContainer>
-                        <UsersChart
-                            userTrends={userTrends}
-                            isLoading={loading}
-                        />
-                    </StyledChartContainer>
-                </StyledWidget>
-            ) : (
-                <StyledWidget>
-                    <StyledWidgetStats>
-                        <WidgetTitle
-                            title={
-                                isOneProjectSelected
-                                    ? 'Users in project'
-                                    : 'Users per project on average'
-                            }
-                            tooltip={
-                                isOneProjectSelected
-                                    ? 'Number of users in selected projects.'
-                                    : 'Average number of users for selected projects.'
-                            }
-                        />
-                        <UserStats
-                            count={summary.averageUsers}
-                            isLoading={loading}
-                        />
-                    </StyledWidgetStats>
-                    <StyledChartContainer>
-                        <UsersPerProjectChart
-                            projectFlagTrends={groupedProjectsData}
-                            isLoading={loading}
-                        />
-                    </StyledChartContainer>
-                </StyledWidget>
-            )}
-        </Section>
     );
 };
