+++ frontend/src/component/insights/sections/LifecycleInsights.tsx
@@ -1,0 +1,26 @@
+import { usePersistentTableState } from 'hooks/usePersistentTableState';
+import type { FC } from 'react';
+import { FilterItemParam } from 'utils/serializeQueryParams';
+import { InsightsSection } from 'component/insights/sections/InsightsSection';
+import { InsightsFilters } from 'component/insights/InsightsFilters';
+
+export const LifecycleInsights: FC = () => {
+    const statePrefix = 'lifecycle-';
+    const stateConfig = {
+        [`${statePrefix}project`]: FilterItemParam,
+    };
+    const [state, setState] = usePersistentTableState('insights', stateConfig);
+
+    return (
+        <InsightsSection
+            title='Flags lifecycle currently'
+            filters={
+                <InsightsFilters
+                    state={state}
+                    onChange={setState}
+                    filterNamePrefix={statePrefix}
+                />
+            }
+        />
+    );
+};
